Move `program/solarflare' into `program/snabbmark'.
