+++ src/program/snabbmark/README
@@ -5,3 +5,11 @@ Usage:
   snabbmark nfvconfig  <pci-address> <config-file-x> <config-file-y> <n>
     Benchmark loading <config-file-y> and transitioning from <config-file-x>
     to <config-file-y> <n> times.
+
+  snabbmark solarflare <npackets> <packet-size>
+    Send the given number of packets through a Solarflare NIC.  The test
+    assumes that the first two Solarflare NICs are connected
+    back-to-back.
+
+    Example usage with 10 million packets, packet size 128 bytes:
+    solarflare 10e6 128
