--- src/program/solarflare/README.inc
@@ -1,1 +1,0 @@
-README.md
